+++ src/main/java/com/benjilebon/saphyre/items/GiftItem.java
@@ -18,20 +18,22 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.storage.loot.LootContext;
 import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSet;
+import net.minecraftforge.common.ForgeConfigSpec;
 import net.minecraftforge.registries.RegistryObject;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class GiftItem extends Item {
     private final String itemName;
     //TODO: custom loot tables ?
-    private final String lootTable;
+    private final ForgeConfigSpec.ConfigValue<String> lootTable;
 
     // We use the registry object entry instead of directly calling the object since it's not yet registered into the particle registry
     private RegistryObject<SimpleParticleType> giftParticle;
 
-    public GiftItem(String itemName, String lootTable, Properties properties)
+    public GiftItem(String itemName, ForgeConfigSpec.ConfigValue<String> lootTable, Properties properties)
     {
         super(properties);
         this.itemName = itemName;
@@ -61,24 +63,34 @@ public class GiftItem extends Item {
             }
             return InteractionResultHolder.success(currentSaphyre);
         } else {
-            LootTable giftTable = ((ServerLevel) level).getServer().getLootTables().get(new ResourceLocation(lootTable));
-            LootContext giftCtx = (new LootContext.Builder((ServerLevel) level)).create(LootContextParamSet.builder().build());
-            List<ItemStack> generatedLoots = giftTable.getRandomItems(giftCtx);
-            ItemStack givenLoot = generatedLoots.get(
-                    ThreadLocalRandom.current().nextInt(generatedLoots.size()) % generatedLoots.size()
-            );
+            try {
+                LootTable giftTable = ((ServerLevel) level).getServer().getLootTables().get(new ResourceLocation(lootTable.get()));
+                if (giftTable.getLootTableId() == null) {
+                    throw new IllegalStateException("Specified loot table for " + itemName + " does not exist (" + lootTable.get() + ")");
+                }
+                LootContext giftCtx = (new LootContext.Builder((ServerLevel) level)).create(LootContextParamSet.builder().build());
+                List<ItemStack> generatedLoots = giftTable.getRandomItems(giftCtx);
+                ItemStack givenLoot = generatedLoots.get(
+                        ThreadLocalRandom.current().nextInt(generatedLoots.size()) % generatedLoots.size()
+                );
 
-            SaphyreMain.LOGGER.debug("Player " + player.getName().getContents() + " used " + this.itemName + " : got " + givenLoot);
-            player.sendSystemMessage(formatSuccessChatMessage(givenLoot));
-            if (this.canGift(player, givenLoot)) {
-                player.addItem(givenLoot);
-            } else {
-                ItemEntity itemEntity = new ItemEntity(level, player.getX(), player.getY() + 1.0D, player.getZ(), givenLoot);
-                level.addFreshEntity(itemEntity);
+                SaphyreMain.LOGGER.debug("Player " + player.getName().getContents() + " used " + this.itemName + "(" + lootTable.get() +")" + " : got " + givenLoot);
+                player.sendSystemMessage(formatSuccessChatMessage(givenLoot));
+                if (this.canGift(player, givenLoot)) {
+                    player.addItem(givenLoot);
+                } else {
+                    ItemEntity itemEntity = new ItemEntity(level, player.getX(), player.getY() + 1.0D, player.getZ(), givenLoot);
+                    level.addFreshEntity(itemEntity);
+                }
+                player.getItemInHand(interactionHand).shrink(1);
+
+                return InteractionResultHolder.pass(player.getItemInHand(interactionHand));
+            } catch (Exception e) {
+                player.sendSystemMessage(Component.translatable("message.gifting.serverError"));
+                SaphyreMain.LOGGER.error(e.getMessage());
+                e.printStackTrace();
+                return InteractionResultHolder.fail(player.getItemInHand(interactionHand));
             }
-            player.getItemInHand(interactionHand).shrink(1);
-
-            return InteractionResultHolder.pass(player.getItemInHand(interactionHand));
         }
     }