+++ src/main/java/com/benjilebon/saphyre/world/features/SaphyrePlacedFeatures.java
@@ -1,7 +1,8 @@
-package com.benjilebon.saphyre.world;
+package com.benjilebon.saphyre.world.features;
 
 import com.benjilebon.saphyre.SaphyreMetadata;
 import com.benjilebon.saphyre.config.SaphyreCommonConfigs;
+import com.benjilebon.saphyre.world.features.SaphyreOreFeatures;
 import net.minecraft.core.Holder;
 import net.minecraft.core.Registry;
 import net.minecraft.world.level.levelgen.VerticalAnchor;
@@ -22,11 +23,14 @@ public class SaphyrePlacedFeatures {
             () -> new PlacedFeature((Holder<ConfiguredFeature<?,?>>)(Holder<? extends ConfiguredFeature<?,?>>)
                     SaphyreOreFeatures.SAPHYRE_FEATURE_ORE,
                     List.of(
-                            CountPlacement.of(SaphyreCommonConfigs.SAPHYRE_ORE_VEINS_PER_CHUNK.get()), // Number of veins in a chunk
+                            // using .get() crashes the game since configs are not loaded yet when registering placed features, so config cannot be used here anymore
+                            // see https://github.com/MinecraftForge/MinecraftForge/pull/8236 and issue #7716 for that matter
+
+                            CountPlacement.of(SaphyreCommonConfigs.SAPHYRE_ORE_VEINS_PER_CHUNK.getDefault()), // Number of veins in a chunk
                             InSquarePlacement.spread(), // Veins distribution method
                             HeightRangePlacement.uniform(
-                                    VerticalAnchor.aboveBottom(SaphyreCommonConfigs.SAPHYRE_ORE_VEINS_MIN_HEIGHT.get()),
-                                    VerticalAnchor.aboveBottom(SaphyreCommonConfigs.SAPHYRE_ORE_VEINS_MAX_HEIGHT.get())
+                                    VerticalAnchor.aboveBottom(SaphyreCommonConfigs.SAPHYRE_ORE_VEINS_MIN_HEIGHT.getDefault()),
+                                    VerticalAnchor.aboveBottom(SaphyreCommonConfigs.SAPHYRE_ORE_VEINS_MAX_HEIGHT.getDefault())
                             ), // see where does aboveBottom starts
                             BiomeFilter.biome() // ?
                     ))